Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
requests
pyjwt
python-multipart

python-dotenv
pytest
